extract index rerender logic to separate file
