+++ packages/frontend/src/components/table/IndexCell.tsx
@@ -6,6 +6,8 @@ interface IndexCellProps {
   className?: string
 }
 
+// the number is updated dynamically inside the client-side code on the bridges page
+// see renderNumbers() inside /packages/frontend/src/scripts/configureCanonicalBridgesFilter.ts
 export function IndexCell({ index, className }: IndexCellProps) {
   return (
     <span
